#7840: Move mistral7b to wormhole folder
