+++ models/demos/wormhole/mistral7b/demo/input_data.json
