+++ models/demos/wormhole/mistral7b/tests/test_mistral_attention.py
@@ -6,14 +6,14 @@ import pytest
 from loguru import logger
 
 import ttnn
-from models.demos.mistral7b.tt.mistral_attention import TtMistralAttention
-from models.demos.mistral7b.tt.mistral_common import (
+from models.demos.wormhole.mistral7b.tt.mistral_attention import TtMistralAttention
+from models.demos.wormhole.mistral7b.tt.mistral_common import (
     precompute_freqs,
     prepare_inputs_ttnn,
     freqs_to_rotation_matrix,
 )
-from models.demos.mistral7b.tt.model_config import TtModelArgs
-from models.demos.mistral7b.reference.model import Attention
+from models.demos.wormhole.mistral7b.tt.model_config import TtModelArgs
+from models.demos.wormhole.mistral7b.reference.model import Attention
 from models.utility_functions import (
     comp_pcc,
     comp_allclose,
