+++ models/demos/wormhole/mistral7b/tests/test_mistral_mlp.py
@@ -6,9 +6,9 @@ import torch
 import pytest
 from loguru import logger
 import ttnn
-from models.demos.mistral7b.tt.model_config import TtModelArgs
-from models.demos.mistral7b.tt.mistral_mlp import TtMistralMLP
-from models.demos.mistral7b.reference.model import FeedForward
+from models.demos.wormhole.mistral7b.tt.model_config import TtModelArgs
+from models.demos.wormhole.mistral7b.tt.mistral_mlp import TtMistralMLP
+from models.demos.wormhole.mistral7b.reference.model import FeedForward
 from models.utility_functions import (
     comp_pcc,
     comp_allclose,
