+++ models/demos/wormhole/mistral7b/tt/generate.py
@@ -4,9 +4,9 @@
 
 from typing import List
 import torch
-from models.demos.mistral7b.reference.model import Transformer, precompute_freqs_cis
-from models.demos.mistral7b.reference.tokenizer import Tokenizer
-from models.demos.mistral7b.tt.model_config import TtModelArgs
+from models.demos.wormhole.mistral7b.reference.model import Transformer, precompute_freqs_cis
+from models.demos.wormhole.mistral7b.reference.tokenizer import Tokenizer
+from models.demos.wormhole.mistral7b.tt.model_config import TtModelArgs
 
 
 def main():
