+++ models/demos/wormhole/mistral7b/tt/mistral_decoder.py
@@ -4,9 +4,9 @@
 import torch
 import ttnn
 from typing import Optional
-from models.demos.mistral7b.tt.mistral_attention import TtMistralAttention
-from models.demos.mistral7b.tt.mistral_mlp import TtMistralMLP
-from models.demos.mistral7b.tt.mistral_rms_norm import TtRMSNorm
+from models.demos.wormhole.mistral7b.tt.mistral_attention import TtMistralAttention
+from models.demos.wormhole.mistral7b.tt.mistral_mlp import TtMistralMLP
+from models.demos.wormhole.mistral7b.tt.mistral_rms_norm import TtRMSNorm
 
 
 class TtTransformerBlock(torch.nn.Module):
